Add the vecSet function and fix bug with vecResize
vecSet sets the data at an index within the vector, overwriting what was
there before (and freeing it). vecSet can also be used to push elements
onto the end of the vector if the index is equal to the leftmost empty
space within the vector.

There was a bug in vecResize where its length was not updated when the
capacity was manually reduced and its elements were truncated. The
vector's length is now updated correctly in these scenarios.